+++ CHANGES.md
@@ -12,6 +12,7 @@
 * Fixed parent not copying when duplicating a style.
 * Fixed unable to export to clipboard with a tree widget in Scene Composer.
 * Fixed background color not being persisted in Scene Composer.
+* Fixed selected drawable not being highlighted in detail view of Drawables dialog.
 * Minor UI tweaks and bug fixes
 
 ### Skin Composer Version 37 ###